refactor: spark application 삭제, spring 의존 주입 field -> 생성자로 변경
