--- src/main/java/chess/dao/spark/DatabaseConnector.java
@@ -1,43 +1,0 @@
-package chess.dao.spark;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-
-public class DatabaseConnector {
-    private final String url;
-    private final String user;
-    private final String password;
-
-    public DatabaseConnector(String url, String user, String password) {
-        this.url = url;
-        this.user = user;
-        this.password = password;
-    }
-
-    public Connection getConnection() throws SQLException {
-        return DriverManager.getConnection(url, user, password);
-    }
-
-    static void close(Connection connection, PreparedStatement statement) {
-        try {
-            statement.close();
-            connection.close();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    static void close(Connection connection, Statement statement, ResultSet resultSet) {
-        try {
-            resultSet.close();
-            statement.close();
-            connection.close();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-}
